More logs

+++ platform-sdk/swirlds-merkledb/src/main/java/com/swirlds/merkledb/collections/LongListOffHeap.java
@@ -152,7 +152,7 @@ public final class LongListOffHeap extends AbstractLongList<ByteBuffer> implemen
                 }
                 if (i == (totalNumOfChunks - 1)) {
                     // last array, so set limit to only the data needed
-                    final long bytesWrittenSoFar = (long) memoryChunkSize * (long) i;
+                    final long bytesWrittenSoFar = (long) memoryChunkSize * i;
                     final long remainingBytes = (size() * Long.BYTES) - bytesWrittenSoFar;
                     buf.limit(toIntExact(remainingBytes));
                 } else {
